refactor(project): namespace indices and caches under RbsInfer::Project (#26)
Half-compact form preserves blame; cross-module references fully
qualified. Require paths updated to project/. CallerFileCache's
ClassMemberCollector ref qualified to RbsInfer:: (its current home)
pending the inference move.
