+++ lib/rbs_infer/project/caller_file_cache.rb
@@ -1,4 +1,4 @@
-module RbsInfer
+module RbsInfer::Project
   # Cache de análise de arquivos caller.
   # ClassMemberCollector, ClassNameExtractor e DefCollector produzem resultados
   # estáveis por arquivo (independem da classe-alvo). Este cache garante que
@@ -28,7 +28,7 @@ module RbsInfer
       comments = result.comments
       lines = entry.source.lines
 
-      member_collector = ClassMemberCollector.new(comments: comments, lines: lines)
+      member_collector = RbsInfer::ClassMemberCollector.new(comments: comments, lines: lines)
       result.value.accept(member_collector)
 
       caller_ext = RbsInfer::AST::ClassNameExtractor.new(file_path: file)
