+++ spec/lib/rbs_infer/source_index_spec.rb
@@ -2,7 +2,7 @@ require "spec_helper"
 require "rbs_infer"
 require "tmpdir"
 
-RSpec.describe RbsInfer::SourceIndex do
+RSpec.describe RbsInfer::Project::SourceIndex do
   around do |example|
     Dir.mktmpdir do |dir|
       @dir = dir
